Added safe closure of testDriver in tearDown method.
In the test class, the safe closure of the testDriver was implemented in the tearDown() method. A try-catch block was added to correctly handle and log any exceptions occurring during the closure which was causing the tests to fail in the Windows environment. This change addresses the bug reported in KAFKA-6647.